Initial implementation of enhanced VDM generation
For details of its inception, see #83. This commit adds an extra
button to the VDM tab that allows segments with an added 'exec'
command before recording.

Lawena now allows the auto-generation (optional) of a .cfg file
called in this exec command, using a given template. This can be
used to automate the workflow of certain batch processes that
involve VDM files.

This commit adds a template to be used with HLAE camera smoothing
and also allows the user to customize the command to execute before
recording a segment.