Workaround Inferno startmovie regression
startmovie command fails with an 'Invalid Path' error unless you
use a path under the tf folder. This makes sure the path is then
resolved as a relative path to work around the regression.

Paths sharing the same root drive as the game also work with this
workaround. Paths in a different drive won't work and no warning
(other than this message) will be given.